Simplify store.js. Don't compose functions when not necessary.
